requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cohere==4.2.1
dataclasses-json==0.5.7
Flask==2.3.1
Flask-Cors==3.0.10
frozenlist==1.3.3
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.54.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.149
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.22.3
pydantic==1.10.7
pypdf==3.8.1
python-dotenv==1.0.0
PyYAML==6.0
qdrant-client==1.1.6
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.10
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.0
yarl==1.9.2